# Optimise lower workloads by not multithreading