… draw-cache/benches/loads-of-unicode.txt