@alexheretic 2021-01-16
Using magnitude `glyph_count * tallest_h * tallest_h` should scale a bit more realistically. This means case `bench_1500_chars_12px` to use single-thread paths, where it was only 24% slower than mt _(I'm looking at roughly 70% slower being the point to select mt)_.

<details><p>

```
group                     mt                    st
-----                     --                    --
bench_1500_chars_150px    1.00     17.0±0.24ms  2.62     44.5±0.23ms
bench_1500_chars_75px     1.00      6.6±0.07ms  2.62     17.2±0.04ms
bench_1500_chars_30px     1.00      3.1±0.09ms  1.97      6.2±0.04ms
bench_1500_chars_12px     1.02      3.0±0.00ms  1.00      2.9±0.02ms
bench_300_chars_150px     1.00      5.1±0.06ms  2.72     14.0±0.18ms
bench_300_chars_75px      1.00  1818.4±26.07µs  2.45      4.5±0.01ms
bench_300_chars_30px      1.00   873.6±46.22µs  1.71  1497.0±16.47µs
bench_300_chars_12px      1.00    765.3±2.30µs  1.00    766.0±5.30µs
bench_50_chars_150px      1.00  1098.7±47.05µs  2.31      2.5±0.00ms
bench_50_chars_75px       1.02    801.3±5.33µs  1.00    788.0±6.46µs
bench_50_chars_30px       1.01    255.4±0.85µs  1.00    252.2±1.05µs
bench_50_chars_12px       1.00    124.2±0.53µs  1.04    128.9±0.74µs
bench_16_chars_150px      1.00    770.4±4.48µs  1.00    768.6±7.30µs
bench_16_chars_75px       1.00    238.7±0.71µs  1.00    239.6±0.92µs
bench_16_chars_30px       1.00     76.7±0.19µs  1.01     77.7±0.31µs
bench_16_chars_12px       1.00     38.4±0.19µs  1.06     40.9±0.17µs
```

</p></details>